optimize Adam and change precision to float to speed up

+++ src/NeuralNetworkGPU/Layer/ConvolutionalLayer.cu
@@ -10,14 +10,14 @@ namespace NeuralNetworkGPU
 	/*
 	 *
 	 */
-	__global__ void determineOutputFuncConv(double *t_input, double *t_output, TensorSize *t_inputSize,
-			double *t_sums,
-			double *t_weights,
+	__global__ void determineOutputFuncConv(float *t_input, float *t_output, TensorSize *t_inputSize,
+			float *t_sums,
+			float *t_weights,
 			MatrixSize *t_filterSize,
-			double *t_deltas,
-			double *d_b)
+			float *t_deltas,
+			float *d_b)
 	{
-		__shared__ double inputBuff[INPUT_BUFFER_SIZE];
+		__shared__ float inputBuff[INPUT_BUFFER_SIZE];
 
 		int inputSize = t_inputSize->x*t_inputSize->y*t_inputSize->z;
 		//copy input to common buffer
@@ -54,7 +54,7 @@ namespace NeuralNetworkGPU
 
 		//sums x[i]*w[i]
 //		long weightsIndex = inputSize*(threadIdx.x + blockIdx.x*blockDim.x);
-		double sum = t_weights[weightsIndex];
+		float sum = t_weights[weightsIndex];
 		for(int i=0; i<inputSize;i++)
 		{
 			sum += inputBuff[i] * t_weights[ weightsIndex+i+1 ];
@@ -69,17 +69,17 @@ namespace NeuralNetworkGPU
 	/*
 	 *
 	 */
-	__global__ void learnBackPropagationFuncConv(double *t_input, int *t_inputSize,
-			double *t_output,
-			double *t_sums,
-			double *t_weights,
-			double *t_deltas, double *t_prevDeltas,
-			double *d_n,double *d_b)
+	__global__ void learnBackPropagationFuncConv(float *t_input, int *t_inputSize,
+			float *t_output,
+			float *t_sums,
+			float *t_weights,
+			float *t_deltas, float *t_prevDeltas,
+			float *d_n,float *d_b)
 	{
 		int inputSize = *t_inputSize;
 
 		//copy input to common buffer
-		__shared__ double inputBuff[INPUT_BUFFER_SIZE];
+		__shared__ float inputBuff[INPUT_BUFFER_SIZE];
 		if(inputSize == blockDim.x)
 		{
 			inputBuff[threadIdx.x] = t_input[threadIdx.x];
@@ -100,15 +100,15 @@ namespace NeuralNetworkGPU
 		__syncthreads();
 
 		long index = threadIdx.x +  blockIdx.x*blockDim.x;
-		double delta = t_deltas[index];
+		float delta = t_deltas[index];
 
 		long weightsIndex = inputSize*(index);
 		//determine common multiplier
-		double e = exp(-(*d_b)*t_sums[index]);
-		double m = 1 + e;
-		double derivative = ((*d_b)*e/(m*m));
+		float e = exp(-(*d_b)*t_sums[index]);
+		float m = 1 + e;
+		float derivative = ((*d_b)*e/(m*m));
 
-		double p = (*d_n)* delta * derivative;
+		float p = (*d_n)* delta * derivative;
 		//calculate new weights
 		//bias weight
 		t_weights[weightsIndex] -= p;
@@ -136,18 +136,18 @@ namespace NeuralNetworkGPU
 	/*
 	 *
 	 */
-	ConvolutionalLayer::ConvolutionalLayer(double t_parameterB, double t_learnRate,
+	ConvolutionalLayer::ConvolutionalLayer(float t_parameterB, float t_learnRate,
 			MatrixSize t_filterSize, TensorSize t_size, TensorSize t_inputSize, NeuronsPtr t_prevLayerReference)
 	{
 		size = t_size;
 		de_input = t_prevLayerReference.inputPtr;
 
 		//learn rate
-		cudaMalloc( (void **) &d_n, sizeof(double));
-		cudaMemcpy(d_n, &(t_learnRate), sizeof(double), cudaMemcpyHostToDevice);
+		cudaMalloc( (void **) &d_n, sizeof(float));
+		cudaMemcpy(d_n, &(t_learnRate), sizeof(float), cudaMemcpyHostToDevice);
 		//parameter b
-		cudaMalloc( (void **) &d_b, sizeof(double));
-		cudaMemcpy(d_b, &(t_parameterB), sizeof(double), cudaMemcpyHostToDevice);
+		cudaMalloc( (void **) &d_b, sizeof(float));
+		cudaMemcpy(d_b, &(t_parameterB), sizeof(float), cudaMemcpyHostToDevice);
 
 		//input size
 		cudaMalloc( (void **) &d_inputSize, sizeof(TensorSize));
@@ -155,21 +155,21 @@ namespace NeuralNetworkGPU
 		inputSize = t_inputSize;
 
 		//output
-		cudaMalloc( (void **) &d_output, sizeof(double)*size.multiply());
-		output = (double*) std::malloc(sizeof(double)*size.multiply());
+		cudaMalloc( (void **) &d_output, sizeof(float)*size.multiply());
+		output = (float*) std::malloc(sizeof(float)*size.multiply());
 
 		//weights
-		cudaMalloc( (void **) &d_weights, sizeof(double)*size.multiply()*t_filterSize.multiply()*t_inputSize.z);
+		cudaMalloc( (void **) &d_weights, sizeof(float)*size.multiply()*t_filterSize.multiply()*t_inputSize.z);
 		initWeights();
 		//filter size
 		cudaMalloc( (void **) &d_filterSize, sizeof(MatrixSize));
 		cudaMemcpy(d_filterSize, &t_filterSize, sizeof(MatrixSize), cudaMemcpyHostToDevice);
 
 		//sums
-		cudaMalloc( (void **) &d_sums, sizeof(double)*size.multiply());
+		cudaMalloc( (void **) &d_sums, sizeof(float)*size.multiply());
 		//deltas
-		cudaMalloc( (void **) &d_deltas, sizeof(double)*size.multiply());
-		deltas = (double*) malloc(sizeof(double)*size.multiply());
+		cudaMalloc( (void **) &d_deltas, sizeof(float)*size.multiply());
+		deltas = (float*) malloc(sizeof(float)*size.multiply());
 		de_prevDeltas = t_prevLayerReference.deltaPtr;
 
 //		numberOfBlocks = 1;
@@ -210,16 +210,16 @@ namespace NeuralNetworkGPU
 	{
 		long weightsSize = inputSize.multiply()*size.multiply()*inputSize.z;
 
-		double *randomValues = (double*) malloc(sizeof(double)*weightsSize);
+		float *randomValues = (float*) malloc(sizeof(float)*weightsSize);
 
 		for(int i=0; i< weightsSize; i++)
 		{
 //			std::cout << (int) (100*i/((inputSize+1)*size)) << "%\n";
-			double randomValue = getRandomWeight();
+			float randomValue = getRandomWeight();
 			randomValues[i] = randomValue;
 
 		}
-		cudaMemcpy(d_weights, randomValues, sizeof(double)*weightsSize, cudaMemcpyHostToDevice);
+		cudaMemcpy(d_weights, randomValues, sizeof(float)*weightsSize, cudaMemcpyHostToDevice);
 		free(randomValues);
 	}
 
@@ -228,7 +228,7 @@ namespace NeuralNetworkGPU
 	 */
 	std::vector<double> ConvolutionalLayer::getOutput()
 	{
-		cudaMemcpy(output, d_output, sizeof(double)*size.multiply(), cudaMemcpyDeviceToHost);
+		cudaMemcpy(output, d_output, sizeof(float)*size.multiply(), cudaMemcpyDeviceToHost);
 
 		std::vector<double> result;
 		int outputSize = size.multiply();
@@ -253,10 +253,10 @@ namespace NeuralNetworkGPU
 		#pragma omp parallel for shared(deltas,size,output, t_z) private(i) default(none)
 		for(int i=0; i<size.multiply(); i++ )
 		{
-			deltas[i] = (double) output[i] - t_z[i];
+			deltas[i] = (float) output[i] - t_z[i];
 		}
 
-		cudaMemcpy(d_deltas, deltas, sizeof(double)*size.multiply(), cudaMemcpyHostToDevice);
+		cudaMemcpy(d_deltas, deltas, sizeof(float)*size.multiply(), cudaMemcpyHostToDevice);
 	}
 
 	void ConvolutionalLayer::learnSGD()
